pitdash/requirements.txt
appengine-python-standard==1.0.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
cachelib==0.9.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
executing==1.2.0
Flask==2.2.3
Flask-Caching==2.0.2
frozendict==2.3.7
google-auth==2.17.2
idna==3.4
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mock==5.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.22.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.15.0
pytz==2023.3
requests==2.28.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
